+++ src/DAL/Curiosity.DAL.EF.Polly/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0] - 2023-01-29
+
+### Changed
+
+- Upgraded `Microsoft`'s packages up to `6.*` versions.
+- Upgraded `Polly` up to `7.2.3`.
+
 ## [1.0.1] - 2021-03-15
 
 ### Added