+++ src/Email/Curiosity.Email.UnisenderGo/CHANGELOG.md
@@ -1,11 +1,17 @@
 # Changelog
 
+## [1.1.0] - 2023-01-29
+
+### Changed
+
+- Upgraded `Newtonsoft.Json` up to `13.0.2`.
+
 ## [1.0.9] - 2023-08-08
 
 ### Fixed
 
 - Upgraded RestSharp to `108.0.3`.
-:wq
+
 ## [1.0.8] - 2022-07-12
 
 ### Fixed