+++ src/Messaging/Curiosity.RabbitMQ/CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 
+## [1.1.0] - 2023-01-29
+
+### Changed
+
+- Upgraded `RabbitMQ.Client` up to `6.4.0`.
+- Upgraded `Newtonsoft.Json` up to `13.0.2`.
+
 ## [1.0.1] - 2022-01-12
-         
+
 ### Added
 
 - Added Rabbit's port to options.