+++ src/RequestProcessing/Curiosity.RequestProcessing.RabbitMQ/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [1.2.3] - 2022-01-30
+
+### Added
+
+- Removed event set from request processing completion action.
+
+## [1.2.2] - 2022-01-30
+
+### Added
+
+- Added `QoS` multiplier.
+
+## [1.2.1] - 2022-01-29
+
+### Fixed
+
+- Added extra event set after RabbitMQ request processing
+
+## [1.2.0] - 2022-01-29
+
+### Changed
+
+- Upgraded dependencies.
+
 ## [1.1.0] - 2022-01-29
 
 ### Changed